Improve login and logout
Add a profile page with the option to logout. Also set the login page as the initial screen so we can't use the app without logging in.

Everything seems to work well except that you can get to the login page if you stand in the BottomTabNavigator and press back button. The login page redirects to the BottomTabNavigator if the user is signed in but in this case nothing happens. Probably because the LoginScreen is not rendered again. Maybe we can disable the back navigation on BottomTabNavigator?