Implement profile/company editing functionality
The StudentProfile and HostProfile components were put into the
profileScreen to easier retrieve the user input. Eventually we would
like to split up the profileScreen into the StudentProfile and
HostProfile components for easier readability.

Some things to do:
- When putting text into the company description the text goes outside
the screen. We would like it to keep putting the text on the line below.
- Atm we can only edit FirstName, LastName and phoneNr. We would like to
be able to add food preferences and a bit more.

Closes #47 #49