Add and implement QRScreen
Builds a basic QR-screen which has the simple functionality to scan
QR-codes of a student-ID and send a companyConnection - request.
Next step is to add a 1-5 rating framework and also implement a QR-code
for the students.

Additional fixes:
Solves problem in profileScreen where ScrollView wouldn't scroll when
dragging on a <TextInput> which has style textAlign: 'center'. Solved
by adding the attribute numberoflines={1} to the <TextInput>.
More on the issue: https://github.com/facebook/react-native/issues/12167