Add higher resolution maps
The map files are separated into the lists of companies and the
actual map. The images are still formatted when loaded, which means
that zooming into an image will show the same resolution. We would like
to fix this in a future release.

Also removed the map image preview in the list of maps.