Dependency issues solved
When using npm it has been a issue that dependencies of the dependencies
are not working as they are outdated or not synced. To both save time
and my will to live I have changed to yarn which fixes this automaticlly
for me. Before in this project yarn has been used, so it shouldnät be an
issue to use it now :)