Darkmode Companies done (90%)
Currently there is a issue where you can't add a specific color more
than grey or white outside of the DropDownPicker which causes it to mix
colors and make the picker ugly. It is possible to create own themes,
however this is something i couldn't solve atm. Currently the picker has
a ebony like color which works currently, but will hopefully be changed
in the future.