Small hotfix: User tickets
Currently the tickets where shown horizontally which causes them to
appear put of screen, not good. Now they are verticle, however the
container is capped to a size which makes the tickets barely visible and
you must scroll inside the container, not good. This is fixed by having
a render specific container height which is based on the amount of
tickets the user have