Bugfix: Filter based on programme
Currently when filtering based on your program we both have the guilds
shortwords and its descriptive title. This causes errors in filtering
due to the fact that the backend/database doesnt use the guild
shortwords values in the DB for filtering and such. Hence when trying to use
the guild names the filtiering is not showing anything. Ths shortwords
are now removed, the PROGRAMME and INDUSTRY variables are now in
alphabetical order and also i removed the ability to choose the guild
shortword in your profile.