Hotfix: Evade crash due to undefined events
In the EventListScreen we filter (delete) events which are lunch or
banquet events due to that they are tickets for events outside of the
fair. When using delete there becomes a hole in the array which causes
it to be undefined. When switched to old events it crashed due to
keySelector cannot read the id as its undefined.